ai-language-tutor/v5/backend/db/supabase_client.py
import os
from supabase import create_client, Client
from dotenv import load_dotenv
from pathlib import Path
import os

# Load environment from root .env
env_path = Path(__file__).parent.parent / ".env"
load_dotenv(env_path)

SUPABASE_URL = os.getenv("SUPABASE_URL")
SUPABASE_KEY = os.getenv("SUPABASE_KEY")
if not SUPABASE_URL or not SUPABASE_KEY:
    raise ValueError("Missing Supabase credentials. Check your .env file.")

# Initialize Supabase client
supabase: Client = create_client(SUPABASE_URL, SUPABASE_KEY)

def get_supabase() -> Client:
    """Dependency override to provide Supabase client"""
    return supabase
