ai-language-tutor/v5/requirements.txt
aioredis
redis==5.2.1
python-docx
fastapi==0.115.11
uvicorn==0.34.0
pydantic==2.10.6
pydantic[email]
aiocache
openai==1.67.0
httpx
python-dotenv
supabase
python-jose[cryptography] 
passlib[bcrypt]
bcrypt<4.0.0
python-multipart 
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0  # Required for TestClient